dealing with cases when image and kernel sizes are similar and opening edge effects
I've made an attempt to have the border effects of anchor methods match
the other methods. This is, in some ways, a reduction in the capabilities
of the anchor method and is generally a bit of a hack.

darcs-hash:20060813112746-fafb9-e9f32e15f87a83e7e20011d98c43891ede3c6f0f.gz