Many threading and line operation issues addressed
There seemed to be bugs showing in vHGW dilate operations along diagonal
directions. This was tracked to problems with reseting histograms. There was
also a subtle problem where a pixel at the end of a line was being skipped.

The anchor opening and closings forming part of a decomposable SE were also
exhibiting problems when threaded. Current fix is to return the double padding
for these operations. I'm not sure if this makes sense or not. It needs more
testing.

darcs-hash:20070619015950-fafb9-acd164a1e8769172050c1ed7084b9653aaa204fb.gz